refactoring

+++ aibench-aibench/src/main/java/es/uvigo/ei/aibench/repository/PluginInstaller.java
@@ -1,23 +1,30 @@
 /*
- * #%L
- * The AIBench basic runtime and plugin engine
- * %%
- * Copyright (C) 2006 - 2017 Daniel Glez-Peña and Florentino Fdez-Riverola
- * %%
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Lesser Public License for more details.
- * 
- * You should have received a copy of the GNU General Lesser Public
- * License along with this program.  If not, see
- * <http://www.gnu.org/licenses/lgpl-3.0.html>.
- * #L%
+Copyright 2007 Daniel Gonzalez Peña, Florentino Fernandez Riverola
+
+
+This file is part of the AIBench Project. 
+
+AIBench Project is free software: you can redistribute it and/or modify
+it under the terms of the GNU Lesser Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+AIBench Project is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser Public License for more details.
+
+You should have received a copy of the GNU Lesser Public License
+along with AIBench Project.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+/*  
+ * PluginFileManager.java
+ *
+ * Created inside the SING research group (http://sing.ei.uvigo.es)
+ * University of Vigo
+ *
+ * Created on 01/04/2009
  */
 package es.uvigo.ei.aibench.repository;
 
@@ -30,10 +37,14 @@ import java.math.BigInteger;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
+import java.util.Set;
 
 import org.apache.log4j.Logger;
 
+import es.uvigo.ei.aibench.Util;
 import es.uvigo.ei.aibench.repository.info.InstallInfo;
 
 /**
@@ -58,6 +69,9 @@ public class PluginInstaller {
 	private final String[] ignoreDirs;
 	private final File pluginsDir;
 	private final File installDir;
+	private final File baseApplicationDir;
+	
+	private Set<String> jarsToReload = new HashSet<String>();
 	
 	private final FileFilter directoryFilter  = new FileFilter() {
 		
@@ -82,15 +96,21 @@ public class PluginInstaller {
 		}
 	};
 	
-	public PluginInstaller(String pluginsDir, String installDir)
+	public PluginInstaller(String pluginsDir, String installDir, String baseApplicationDir)
 	throws IllegalArgumentException {
-		this(pluginsDir, installDir, null);
+		this(pluginsDir, installDir, baseApplicationDir, null);
 	}
 
-	public PluginInstaller(String pluginsDir, String installDir, String[] ignoreDirs) 
+	public PluginInstaller(String pluginsDir, String installDir,String baseApplicationDir, String[] ignoreDirs) 
 	throws IllegalArgumentException {
 		this.pluginsDir = new File(pluginsDir);
+		
 		this.installDir = new File(installDir);
+		
+		String aux = baseApplicationDir;
+		if(aux == null)
+			aux = ".";
+		this.baseApplicationDir = new File(aux);
 		this.ignoreDirs = ignoreDirs;
 		
 		if (!this.pluginsDir.isDirectory() && !this.pluginsDir.mkdir())
@@ -163,7 +183,7 @@ public class PluginInstaller {
 					return PluginInstaller.checkMD5(info.md5, pluginFile);
 				}
 			} catch (IOException e) {
-				PluginInstaller.logger.error("Error validating install files: " + uid, e);
+//				PluginInstaller.logger.error("Error validating install files: " + uid, e);
 				return false;
 			} catch (NoSuchAlgorithmException e) {
 				PluginInstaller.logger.error("Error validating md5 checksum for plugin install files: " + uid, e);
@@ -199,11 +219,55 @@ public class PluginInstaller {
 	}
 	
 	public void installPlugins(boolean deleteInvalidUpdates) {
+////		boolean updateClassPath = false;
+//		
+//		for (String uid:this.getInstallUIDs()) {
+//			PluginInstaller.logger.info("Installing plugin: " + uid);
+//			if (this.isValidInstall(uid)) {
+//				try {
+//					if (this.installPlugin(uid)) {
+//						PluginInstaller.logger.info("Plugin installed: " + uid);
+//					} else {
+//						PluginInstaller.logger.error("Plugin can't be installed: " + uid);
+//						if (deleteInvalidUpdates) {
+//							PluginInstaller.logger.error(
+//								String.format("Error installing plugin %s. The plugin install files will be deleted.", uid)
+//							);
+//							this.deleteInstall(uid);
+//						}
+//					}
+//				} catch (IOException ioe) {
+//					if (deleteInvalidUpdates) {
+//						PluginInstaller.logger.error(
+//							String.format("Error installing plugin %s. The plugin install files will be deleted.", uid),
+//							ioe
+//						);
+//						this.deleteInstall(uid);
+//					} else {
+//						PluginInstaller.logger.error("Error installing plugin: " + uid, ioe);
+//					}
+//				}
+//			} else {
+//				PluginInstaller.logger.warn("Install files not valid: " + uid);
+//				if (deleteInvalidUpdates) 
+//					this.deleteInstall(uid);
+//			}
+//		}
+//		
+//		if(jarsToReload.size()>0){
+//			Util.reloadClassPath(jarsToReload);
+//		}
+		
+		this.installPlugins(deleteInvalidUpdates, true);
+	}
+	
+	public void installPlugins(boolean deleteInvalidUpdates, boolean delete) {
+//		boolean updateClassPath = false;
 		for (String uid:this.getInstallUIDs()) {
 			PluginInstaller.logger.info("Installing plugin: " + uid);
 			if (this.isValidInstall(uid)) {
 				try {
-					if (this.installPlugin(uid)) {
+					if (this.installPlugin(uid, delete)) {
 						PluginInstaller.logger.info("Plugin installed: " + uid);
 					} else {
 						PluginInstaller.logger.error("Plugin can't be installed: " + uid);
@@ -231,18 +295,99 @@ public class PluginInstaller {
 					this.deleteInstall(uid);
 			}
 		}
+		
+		if(jarsToReload.size()>0){
+			Util.reloadClassPath(jarsToReload);
+		}
 	}
 	
+
 	public boolean installPlugin(String uid)
+	throws IOException {
+//		File uidInstallDir = new File(this.installDir, uid);
+//		
+//		InstallInfo info = new InstallInfo(uidInstallDir);
+//		
+//		installDeps(info, uidInstallDir);
+//		
+//		if (info.file.endsWith("jar")) {
+//			File jarFile = new File(uidInstallDir, info.file);
+//			if (info.updatePlugin == null) {
+//				if (jarFile.renameTo(new File(this.pluginsDir, info.file))) {
+//					PluginInstaller.deleteFile(uidInstallDir);
+//					return true;
+//				} else {
+//					return false;
+//				}
+//			} else {
+//				File updatePluginFile = new File(this.pluginsDir, info.updatePlugin);
+//				File updatePluginOld = new File(
+//					this.pluginsDir, 
+//					String.format(PluginInstaller.OLD_FILE_FORMAT, info.updatePlugin)
+//				);
+//				if (updatePluginFile.renameTo(updatePluginOld) 
+//					&& jarFile.renameTo(new File(this.pluginsDir, info.file))) {
+//					PluginInstaller.deleteFile(uidInstallDir);
+//					PluginInstaller.deleteFile(updatePluginOld);
+//					return true;
+//				} else {
+//					return false;
+//				}
+//			}
+//		} else {
+//			File installFile = new File(uidInstallDir, info.file);
+//			if (info.updatePlugin == null) {
+//				File pluginDir = new File(this.pluginsDir, uid);
+//				File outputDir = new File(
+//					this.pluginsDir, 
+//					String.format(PluginInstaller.INSTALL_FILE_FORMAT, uid)
+//				);
+//				if (Inflater.inflate(installFile, outputDir)!=null &&
+//					outputDir.renameTo(pluginDir)) {
+//					PluginInstaller.deleteFile(uidInstallDir);
+//					return true;
+//				} else {
+//					return false;
+//				}
+//			} else {
+//				File updatePluginFile = new File(this.pluginsDir, info.updatePlugin);
+//				File pluginDir = new File(this.pluginsDir, uid);
+//				File updatePluginOld = new File(
+//					this.pluginsDir, 
+//					String.format(PluginInstaller.OLD_FILE_FORMAT, info.updatePlugin)
+//				);
+//				File outputDir = new File(
+//					this.pluginsDir, 
+//					String.format(PluginInstaller.INSTALL_FILE_FORMAT, uid)
+//				);
+//				if (updatePluginFile.renameTo(updatePluginOld) 
+//					&& Inflater.inflate(installFile, outputDir)!=null 
+//					&& outputDir.renameTo(pluginDir)) {
+//					PluginInstaller.deleteFile(uidInstallDir);
+//					PluginInstaller.deleteFile(updatePluginOld);
+//					return true;
+//				} else {
+//					return false;
+//				}
+//			}
+//		}
+		
+		return this.installPlugin(uid, true);
+	}
+	
+	public boolean installPlugin(String uid, boolean delete)
 	throws IOException {
 		File uidInstallDir = new File(this.installDir, uid);
 		
 		InstallInfo info = new InstallInfo(uidInstallDir);
+		
+		installDeps(info, uidInstallDir);
 		if (info.file.endsWith("jar")) {
 			File jarFile = new File(uidInstallDir, info.file);
 			if (info.updatePlugin == null) {
 				if (jarFile.renameTo(new File(this.pluginsDir, info.file))) {
-					PluginInstaller.deleteFile(uidInstallDir);
+					if(delete)
+						PluginInstaller.deleteFile(uidInstallDir);
 					return true;
 				} else {
 					return false;
@@ -253,16 +398,22 @@ public class PluginInstaller {
 					this.pluginsDir, 
 					String.format(PluginInstaller.OLD_FILE_FORMAT, info.updatePlugin)
 				);
-				if (updatePluginFile.renameTo(updatePluginOld) 
-					&& jarFile.renameTo(new File(this.pluginsDir, info.file))) {
-					PluginInstaller.deleteFile(uidInstallDir);
-					PluginInstaller.deleteFile(updatePluginOld);
-					return true;
-				} else {
+				
+				if (updatePluginFile.renameTo(updatePluginOld)){
+					boolean b = false;
+					if (jarFile.renameTo(new File(this.pluginsDir, info.file))) {
+						if(delete)
+							PluginInstaller.deleteFile(uidInstallDir);
+						PluginInstaller.deleteFile(updatePluginOld);
+						b=true;
+					}else
+						updatePluginFile.renameTo(new File(this.pluginsDir, info.updatePlugin));
+					
+					return b;
+				} else
 					return false;
-				}
 			}
-		} else {
+		}  else {
 			File installFile = new File(uidInstallDir, info.file);
 			if (info.updatePlugin == null) {
 				File pluginDir = new File(this.pluginsDir, uid);
@@ -270,9 +421,10 @@ public class PluginInstaller {
 					this.pluginsDir, 
 					String.format(PluginInstaller.INSTALL_FILE_FORMAT, uid)
 				);
-				if (Inflater.inflate(installFile, outputDir) &&
+				if (Inflater.inflate(installFile, outputDir)!=null &&
 					outputDir.renameTo(pluginDir)) {
-					PluginInstaller.deleteFile(uidInstallDir);
+					if(delete)
+						PluginInstaller.deleteFile(uidInstallDir);
 					return true;
 				} else {
 					return false;
@@ -289,18 +441,52 @@ public class PluginInstaller {
 					String.format(PluginInstaller.INSTALL_FILE_FORMAT, uid)
 				);
 				if (updatePluginFile.renameTo(updatePluginOld) 
-					&& Inflater.inflate(installFile, outputDir) 
+					&& Inflater.inflate(installFile, outputDir)!=null 
 					&& outputDir.renameTo(pluginDir)) {
-					PluginInstaller.deleteFile(uidInstallDir);
+					if(delete)
+						PluginInstaller.deleteFile(uidInstallDir);
+					
 					PluginInstaller.deleteFile(updatePluginOld);
 					return true;
-				} else {
+				} else
 					return false;
-				}
 			}
 		}
 	}
 	
+	/**
+	 * @param info
+	 * @param  
+	 */
+	private void installDeps(InstallInfo info, File uidInstallDir) {
+		
+		Iterator<String> folders = info.depFolders.iterator();
+		for(String dep : info.dependencies){
+			String folder = folders.next();
+			
+//			System.out.println("\n\nInfalter");
+//			System.out.println("Dependencie: " + dep);
+//			System.out.println("Folder: " + folder);
+//			System.out.println("InstallDir: " + this.installDir);
+//			System.out.println("baseApplicationDir: " + this.baseApplicationDir);
+			
+			File output =  new File(this.baseApplicationDir,folder);
+			File depFile =  new File(uidInstallDir,dep);
+			
+			
+//			System.out.println("out: " + output.getAbsolutePath()+ "\t" +output.exists());
+//			System.out.println("depFile: " + depFile.getAbsolutePath() + "\t" +depFile.exists());
+			
+			Set<String> files = Inflater.inflate(depFile,output);
+//			System.out.println(files);
+			
+			if(files!=null)
+			jarsToReload.addAll(files);
+			
+			
+		}
+	}
+
 	private static boolean checkMD5(String testMD5, File testFile)
 	throws IOException, NoSuchAlgorithmException {
 		BufferedInputStream bis = null;
